util: define memory pool utility routines
A memory pool consists of a preallocated portion and a dynamic allocated
portion. The preallocated portion is allocated by the calling routine
and initialized when the first alloc call is made. The dynamic portion
is created when the preallocated objects are used up. The dynamic
portion may contain multiple memory blocks. Each time it reallocates
a memory block which contains multiple objects. All blocks are freed
when the memory pool is destroyed.

The memory pool utility is used for collective communicator cache and
strided datatype cache.