sobj: define symm object handle for AM
We define sobj handle for each symm segment type (sheap, sdata,space
heap) and transfer it to the target PE when issuing AM based operations.
The target PE calculates the absolute vaddr of dest based on the handle.

Note: this is the first refactoring patch for adding space heap. The
code will be significantly changed again by a later commit